refactor(editor): 섹션 스트립 로직 분리 및 미사용 컴포넌트 정리
