+++ troubleshooting.md
@@ -1,0 +1,5 @@
+# Troubleshooting Log
+
+| Date       | Type      | Description           | Resolution                          |
+| ---------- | --------- | --------------------- | ----------------------------------- |
+| 2025-12-24 | AI Review | PR #3 AI Review Check | ✅ Pass - No critical issues found. |
